requirements.txt
aiohttp==3.7.4.post0
discord.py==1.7.3
jishaku==2.0.0
psutil==5.8.0
beautifulsoup4==4.9.3
mintchoco==4.0.4
sentry-sdk==1.1.0
